Big updates to the Live route.
Added a PlayView which is used while the question is asked and the timer is running.
    Basically, it acts as the place where the question and the answers are held/displayed

Added a MultipleChoice component to be used inside PlayView. Its purpose is to handle multiple choice questions.
    It is the holder of the choices, and also handles actions all the way up until submission.

Added InfoView which shows information regarding the active game.

Updated TeamView and WaitView to include a style prop, and also added info to when waiting for the question, while the question is running.

ActivityStream now is clickable to return to PlayView

I also added opacity animations to Views changing (ie. when the user switches Views)